# api/openapi.yaml
openapi: 3.1.0
info:
  title: Agentic Ecommerce Mission Control API
  version: 9.0.0
  description: |
    Public v1.x.y contract for the Agentic Ecommerce Go backend consumed by
    agentic-ecommerce-web. This spec intentionally excludes private
    mobile-only operator endpoints and credential-bearing surfaces.
    Protected endpoints use short-lived HMAC-signed JWT bearer tokens and RBAC roles:
    admin, operator, and viewer.

    Version policy: v1 endpoints are stable through host v9.x. The
    v2 preview surface (api/openapi-v2-preview.yaml) is opt-in only
    and subject to change without notice. See docs/api-versioning.md
    for negotiation, deprecation timeline, and Accept-header opt-in.
  x-api-version-policy: |
    v1 endpoints are stable through host v9.x; v2 preview endpoints are
    subject to change without notice. See docs/api-versioning.md.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/workflows/product-publish:
    post:
      summary: Start the ProductPublishWorkflow
      operationId: startProductPublishWorkflow
      description: Starts the canonical Temporal workflow that checks compliance, validates media, waits for human review, and publishes to WooCommerce.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartProductPublishWorkflowRequest'
      responses:
        '202':
          description: Temporal workflow execution accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowStartResponse'
        '400':
          description: Invalid JSON body or product ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Temporal client is not configured for the API process.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Temporal rejected the workflow start request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/workflows/marketplace-sync:
    post:
      summary: Start the MarketplaceSyncWorkflow
      operationId: startMarketplaceSyncWorkflow
      description: Starts the Temporal workflow that dispatches a product sync event to the configured marketplace provider router.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartMarketplaceSyncWorkflowRequest'
      responses:
        '202':
          description: Temporal workflow execution accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowStartResponse'
        '400':
          description: Invalid JSON body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Required marketplace sync fields are missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Temporal client is not configured for the API process.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Temporal rejected the workflow start request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/workflows/marketplace-replay:
    post:
      summary: Start the MarketplaceReplayWorkflow
      operationId: startMarketplaceReplayWorkflow
      description: Starts the Temporal workflow that replays one marketplace DLQ record through the provider router.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartMarketplaceReplayWorkflowRequest'
      responses:
        '202':
          description: Temporal workflow execution accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowStartResponse'
        '400':
          description: Invalid JSON body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Required marketplace replay fields are missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Temporal client is not configured for the API process.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Temporal rejected the workflow start request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/workflows/content-generation:
    post:
      summary: Start the ContentGenerationWorkflow
      operationId: startContentGenerationWorkflow
      description: Starts the Temporal workflow that generates content, fact-checks claims against RAG evidence, evaluates quality, and auto-approves or rejects the result.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartContentGenerationWorkflowRequest'
      responses:
        '202':
          description: Temporal workflow execution accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowStartResponse'
        '400':
          description: Invalid JSON body or product ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Temporal client is not configured for the API process.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Temporal rejected the workflow start request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/workflows/media-processing:
    post:
      summary: Start the MediaProcessingWorkflow
      operationId: startMediaProcessingWorkflow
      description: Starts the Temporal workflow that sources supplier media, processes it, runs QA, stores it, and links it to a product.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartMediaProcessingWorkflowRequest'
      responses:
        '202':
          description: Temporal workflow execution accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowStartResponse'
        '400':
          description: Invalid JSON body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Product ID or source URL is missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Temporal client is not configured for the API process.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Temporal rejected the workflow start request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/workflows/sourcing:
    post:
      summary: Start the SourcingWorkflow
      operationId: startSourcingWorkflow
      description: Starts the Temporal workflow that searches deterministic supplier candidates, scores them, compares landed prices, checks margin, and recommends a supplier.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartSourcingWorkflowRequest'
      responses:
        '202':
          description: Temporal workflow execution accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowStartResponse'
        '400':
          description: Invalid JSON body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: SKU or deterministic candidate list is missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Temporal client is not configured for the API process.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Temporal rejected the workflow start request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/workflows:
    get:
      summary: List Temporal workflows with lifecycle visibility
      operationId: listWorkflows
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, running, waiting_review, completed, failed, canceled, terminated, continued_as_new, timed_out, unspecified]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Workflow lifecycle summaries for operator and UI surfaces.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowListResponse'
        '503':
          description: Temporal client is not configured for the API process.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Temporal visibility request failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/workflows/{id}:
    get:
      summary: Get Temporal workflow lifecycle detail
      operationId: getWorkflowStatus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Temporal workflow execution detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDetailResponse'
        '404':
          description: Workflow execution was not found in Temporal visibility.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Temporal client is not configured for the API process.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Temporal describe request failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/workflows/{id}/signals/review:
    post:
      summary: Signal ProductPublishWorkflow human review
      operationId: signalProductPublishReview
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductPublishReviewSignal'
      responses:
        '202':
          description: Review signal accepted by Temporal with the latest observed workflow snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowSignalResponse'
        '400':
          description: Invalid JSON body or workflow ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Temporal client is not configured for the API process.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Temporal signal request failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HMAC-signed JWT access token issued by `/api/v1/auth/login` or `/api/v1/auth/refresh`.
  schemas:
...
    StartProductPublishWorkflowRequest:
      type: object
      required: [product_id]
      properties:
        product_id:
          type: string
          format: uuid
        requested_by:
          type: string
          description: Operator identity recorded in workflow audit events.
...
    WorkflowStartResponse:
      type: object
      required: [workflow_id, run_id, status, task_queue]
      properties:
        workflow_id:
          type: string
        run_id:
          type: string
        status:
          type: string
          enum: [started]
        task_queue:
          type: string
          enum: [ec-workflows]
...
    StartMarketplaceSyncWorkflowRequest:
      type: object
      required: [event]
      properties:
        event:
          $ref: '#/components/schemas/MarketplaceSyncEvent'
    StartMarketplaceReplayWorkflowRequest:
      type: object
      required: [record]
      properties:
        record:
          $ref: '#/components/schemas/MarketplaceDLQRecord'
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...